fix: separate delete with validation of server
