+++ app/Http/Livewire/Server/Form.php
@@ -4,12 +4,10 @@ namespace App\Http\Livewire\Server;
 
 use App\Actions\Server\InstallDocker;
 use App\Models\Server;
-use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Livewire\Component;
 
 class Form extends Component
 {
-    use AuthorizesRequests;
     public Server $server;
     public bool $isValidConnection = false;
     public bool $isValidDocker = false;
@@ -106,23 +104,9 @@ class Form extends Component
         }
     }
 
-    public function delete()
-    {
-        try {
-            $this->authorize('delete', $this->server);
-            if (!$this->server->isEmpty()) {
-                $this->emit('error', 'Server has defined resources. Please delete them first.');
-                return;
-            }
-            $this->server->delete();
-            return redirect()->route('server.all');
-        } catch (\Throwable $e) {
-            return handleError($e, $this);
-        }
-    }
     public function submit()
     {
-        if(isCloud() && !isDev()) {
+        if (isCloud() && !isDev()) {
             $this->validate();
             $this->validate([
                 'server.ip' => 'required',
